mirror-pack-agent/requirements.txt
PyYAML>=6.0.3
jsonschema>=4.23.0
